+++ CHANGELOG.md
@@ -67,6 +67,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added ellipsis to too long links (#274)
 -   New logo (#281)
 -   Align TaskDrawer children to the left (#305)
+-   Change Download button appearance (#308)
 
 ## [0.6.0] - 2021-12-01